wizard_requirements.txt
streamlit
requests
python-dotenv
langchain
langchain-openai
